[Fix][Transform-V2] Fix multiTable SQL transform
